// yanay-tevet-web/src/generated-files/api/blocks/types.gen.ts
export type BlockTypes = 'round' | 'square' | 'triangle';
// ...
export type BlockSchema = {
    /**
     * Id
     */
    id: number;
    /**
     * A
     */
    a: string;
    /**
     * B
     */
    b: number;
    /**
     * C
     */
    c: boolean;
    block_type: BlockTypes;
};
// ...
export type PaginationBlockViewData = {
    body?: never;
    path?: never;
    query?: {
        /**
         * Page
         */
        page?: number;
        /**
         * Page Size
         */
        page_size?: number;
        /**
         * Order By
         */
        order_by?: Array<string> | null;
        /**
         * Dict Filter
         */
        dict_filter?: string | null;
        /**
         * A
         */
        a?: string | null;
        /**
         * C
         */
        c?: string | null;
        /**
         * Search
         */
        search?: string | null;
        block_type?: BlockTypes | null;
    };
    url: '/api/blocks/';
};